Fixed leveldown-mobile failing to install on Windows.
Fixes https://github.com/thaliproject/Thali_CordovaPlugin/issues/763
The node-gyp version that comes with npm 3.3.12 is incompatible with
Visual Studio 2015.
The fix is the same fix applied for npm 3.4.0
https://github.com/nodejs/node-gyp/commit/f31482e2260e3beb90d959c62135e5e95c06ed4d
In order to have this fix available after merging it, npmjx311.tar.gz
needs to be recreated using deps/packnpm.sh and uploaded to:
https://jxcore.azureedge.net/npmjx/npmjx311.tar.gz
Be aware that on a clean machine packnpm.sh produces a npmjx311.tar.gz
file that is different from the current version at
https://jxcore.azureedge.net/npmjx/npmjx311.tar.gz
In particular I noticed on my machine that it's missing the 'debug'
submodule for:
npm/node_modules/node-gyp/node_modules/path-array/node_modules/array-index
that was causing an error when installing leveldown-mobile.
I manually executed 'jx npm install debug' to fix it.
I will have to investigate it further.