Remove explicit GC calls in jxcore.
There is a bug that causes the SpiderMonkey's GC to crash. This change is not
a fix for that bug, but explicit calls to GC are worsening the bug effects.
Once the underline bug will be fixed we will reconsider adding back the
explicit GC calls.
It's also important to stress that 'gc()' calls in the javascript code are
worsening the effects of the bug and should be avoided.